+++ permission/README.md
@@ -1,0 +1,2 @@
+# permissionutils
+The package contains various helpers about permissions/privileges
