+++ permission/error.go
@@ -1,0 +1,5 @@
+package permissionutil
+
+import "errors"
+
+var ErrNotImplemented = errors.New("not implemented")
