+++ permission/permission.go
@@ -1,0 +1,11 @@
+package permissionutil
+
+var (
+	IsRoot       bool
+	HasCapNetRaw bool
+)
+
+func init() {
+	IsRoot, _ = checkCurrentUserRoot()
+	HasCapNetRaw, _ = checkCurrentUserCapNetRaw()
+}
