--- privileges/README.md
@@ -1,2 +1,0 @@
-# privilegesutil
-The package contains various helpers for previleges
