--- privileges/privileges.go
@@ -1,7 +1,0 @@
-package privileges
-
-var IsPrivileged bool
-
-func init() {
-	IsPrivileged = isPrivileged()
-}
