--- privileges/privileges_win.go
@@ -1,8 +1,0 @@
-//go:build windows
-
-package privileges
-
-// IsPrivileged on windows doesn't matter as we are using connect scan
-func isPrivileged() bool {
-	return false
-}
